Refactor utils files and update import paths
